`basestring` variable not used in Python 3, add compatibility layer
Since all strings are Unicode in Python 3, the variables `unicode` and `basestring` are no longer used. This patch makes these variables work again in Python 3.

http://www.rfk.id.au/blog/entry/preparing-pyenchant-for-python-3/

However, the real solution is to modify `get_boards()` such that it no longer needs to use `basestring`. Though it has to be done in a way that both Python 2 and 3 will accept.